Move generateAddress script to utils folder
